restructured trip component
