+++ frontend/src/components/TripList/Trip/DestinationList/index.jsx
