+++ frontend/src/components/TripList/index.jsx
@@ -1,0 +1,12 @@
+import React from "react";
+import Trip from "./Trip";
+
+function TripList() {
+    return (
+        <div>
+            <Trip />
+        </div>
+    );
+}
+
+export default TripList;
